RABSW-927: Include compute information in directive breakdown (#11)
* RABSW-927: Include compute information in directive breakdown

This commit changes the DirectiveBreakdown resource to include both
storage and compute information. The storage information describes the
storage requirements and how they should be allocated. The compute
information describes the compute requirements and how they should be
picked.

I also moved some fields from the Spec to the Status since the
DirectiveBreakdown controller fills them in based on the directive provided
in the spec. I also removed some fields that aren't necessary for the
WLM to know about (e.g., FS type and DWD index).

Signed-off-by: Matt Richerson <mattr@cray.com>

* DirectiveBreakdownSpec comment change

Signed-off-by: Matt Richerson <mattr@cray.com>

Co-authored-by: Matt Richerson <mattr@cray.com>