RABSW-996: Handle multiple #DW errors properly
This commit improves how the DWS workflow controller handles multiple #DWs
that have registered for the same workflow phases. If any of the driverStatus
elements for the current phase have an error, then one of the errors is reported
in the workflow status. If that error is resolved, then the workflow status shows
any other errors fom other driverStatus elements. Once there are no errors, then
the workflow status no longer shows an error.

Signed-off-by: Matt Richerson <mattr@cray.com>