RABSW-999: Add error field to driver status
This commit adds and error field to the entries in thedriverstatus array.
The error field is meant to provide a non-user facing error message when
something goes wrong. This field is not meant to be rolled up into the overall
status of the workflow, and it is not parsed by the WLM. This field is intended
to be used by admins and developers.

Signed-off-by: Matt Richerson <mattr@cray.com>