Revert to Ginkgo V1.
Kubebuilder does not yet support Ginkgo v2, and sometimes craps out when we
ask it to do new things in this repo when v2 is in use.

Upgrade controller-gen from 0.6.1 to 0.9.2, which causes some annotations to
be updated in the CRDs.

Update Go from 1.16 to 1.17 in the Dockerfile to address an error between
sigs.k8s.io/json and reflect.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>