Add UserID to DirectiveBreakdown and PersistentStorageInstance (#24)
This commit adds the UserID field to the DirectiveBreakdown and PersistentStorageInstance
to track which user created a persistent storage and to verify that two users don't successfully
create the same persistent storage.

This commit also adds a SetResourceError() function that can be called when a reconciler is finished
to fill in object.Status.Error with a information from the returned error.

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>