Add UserID to DirectiveBreakdown and PersistentStorageInstance (#24)
This commit adds the UserID field to the DirectiveBreakdown and PersistentStorageInstance
to track which user created a persistent storage and to verify that two users don't successfully
create the same persistent storage.

This commit also adds an embedded ResourceError struct in the Status section of the resources
to provide a common receiver function for setting the error information.

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>