Support CSM.
Use the DP0 overlay (real hardware) for the config base.  CSM currently uses
an ancient cert-manager at v0.14.1, so set the apiVersion back to v1alpha3 to
accomodate that.

The CSM systems we are testing on do not have access to gcr.io, so we have to
use skopeo to copy kube-rbac-proxy into the CSM system's own container
registry.  Patch the Deployment to find it in this location.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>