RABSW-1096: Add constraints to DirectiveBreakdown for Lustre targets
Add two new constraints to the storage allocations in the DirectiveBreakdown that will
be used to determine how many Lustre targets to create. The two new constraints are:

- Count: A simple count field the specifies exactly how many allocations to make
- Scale: A 1-10 value that is a magnitude of how many allocations to make.

Also change the computes location constraint to allow multiple connection types and
specify whether the connection is mandatory. This allows Lustre targets to specify a
"BestEffort" physical connection to their computes to try to keep compute nodes and
Lustre targets in the same chassis. The "BestEffort" priority is used because the
number of compute nodes and number of Rabbits used may not match well.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>