Limit the SystemConfiguration CRD conversion
When converting to the v1alpha1 spoke do not add the annotation that is
the serialized version of the v1alpha2 hub. On a customer site that annotation
is too big for the k8s limit of 262144 bytes.

With this change, we can still convert v1alpha2->v1alpha1 and v1alpha1->v1alph2,
but we can't convert v1alpha1->v1alpha2->v1alpha1 and
v1alpha2->v1alpha1->v1alpha2. However those last two (1-2-1 and 2-1-2) happen
only in the conversion test suite, and that is disabled now.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>