Add RequiredDaemons list to DirectiveBreakdown
Allow a user to specify a list of required daemons to their "#DW jobdw ..."
directive. The new keyword is "requires=" and it takes a list of strings:

  requires=copy-offload,future-thing

The dwdparse library is updated to support parsing a comma-separated list
of strings, with a list of RE patterns for each element allowed in the list.

The DirectiveBreakdownStatus has a new field RequiredDaemons to hold the list
of daemons.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>