Add SystemStatus resource
This commit adds a new resource that is used to allow an external component (e.g., Flux), to
set the status of nodes on the system. If the node is listed and its status is "Disabled",
then the Rabbit software will consider the node down. If the node isn't listed, its status
is assumed to be "Enabled"

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>