CRDBUMPER-bump-apis
Point earlier spoke APIs at new hub v1alpha6.

The conversion_test.go and the ConvertTo()/ConvertFrom() routines in
conversion.go are still valid for the new hub because it is currently
identical to the previous hub.

Update the k8s:conversion-gen marker in doc.go to point to the new hub.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>